+++ mobile/shared/components/geckoview/GeckoViewStartup.sys.mjs
@@ -218,6 +218,7 @@ export class GeckoViewStartup {
               "GeckoView:IPProtection:Activate",
               "GeckoView:IPProtection:Deactivate",
               "GeckoView:IPProtection:Enroll",
+              "GeckoView:IPProtection:RefreshUsage",
             ],
           });